feat: add / update logs
