+++ connectors-sdk/connectors_sdk/settings/base_settings.py
@@ -6,12 +6,11 @@ to manage and validate configuration parameters using Pydantic.
 These models can be extended to create specific configurations for different types of connectors.
 """
 
-import sys
+from __future__ import annotations
+
 from abc import ABC
-from copy import deepcopy
 from datetime import timedelta
-from pathlib import Path
-from typing import Any, ClassVar, Literal, Self
+from typing import TYPE_CHECKING, Any, ClassVar, Literal, Self
 
 from connectors_sdk.logging.logger import Logger
 from connectors_sdk.logging.sdk_logger import sdk_logger
@@ -33,17 +32,12 @@ from pydantic import (
     HttpUrl,
     ModelWrapValidatorHandler,
     ValidationError,
-    create_model,
     model_validator,
 )
 from pydantic.fields import FieldInfo
-from pydantic_settings import (
-    BaseSettings,
-    DotEnvSettingsSource,
-    PydanticBaseSettingsSource,
-    SettingsConfigDict,
-    YamlConfigSettingsSource,
-)
+
+if TYPE_CHECKING:
+    from connectors_sdk.logging._base_logger import BaseLogger
 
 
 class BaseConfigModel(BaseModel, ABC):
@@ -124,137 +118,82 @@ class _BaseConnectorConfig(BaseConfigModel, ABC):
     )
 
 
-class _SettingsLoader(BaseSettings):
-    model_config = SettingsConfigDict(
-        frozen=True,
-        extra="allow",
-        env_nested_delimiter="_",
-        env_nested_max_split=1,
-        enable_decoding=False,
+class BaseExternalImportConnectorConfig(_BaseConnectorConfig):
+    """Settings class for external import connectors.
+
+    Attributes:
+        type (str): The type of the connector, set to "EXTERNAL_IMPORT" for external import connectors.
+        duration_period (timedelta): The period of time to await between two runs of the connector.
+    """
+
+    type: Literal["EXTERNAL_IMPORT"] = "EXTERNAL_IMPORT"
+    duration_period: timedelta = Field(
+        description="The period of time to await between two runs of the connector."
     )
 
-    @classmethod
-    def _get_connector_main_path(cls) -> Path:
-        """Locate the main module of the running connector.
-        This method is used to locate configuration files relative to connector's entrypoint.
 
-        Notes:
-            - This method assumes that the connector is launched using a file-backed entrypoint
-            (i.e., `python -m <module>` or `python <file>`).
-            - At module import time, `__main__.__file__` might not be available yet,
-            thus this method should be called at runtime only.
-        """
-        main = sys.modules.get("__main__")
-        if main and getattr(main, "__file__", None):
-            return Path(main.__file__).resolve()  # type: ignore
+class BaseInternalEnrichmentConnectorConfig(_BaseConnectorConfig):
+    """Settings class for internal enrichment connectors.
 
-        raise RuntimeError(
-            "Cannot determine connector's location: __main__.__file__ is not available. "
-            "Ensure the connector is launched using `python -m <module>` or a file-backed entrypoint."
-        )
+    Attributes:
+        type (str): The type of the connector, set to "INTERNAL_ENRICHMENT" for internal enrichment connectors.
+        auto (bool): Whether the connector should run automatically when an entity is created or updated.
+    """
 
-    @classmethod
-    def _get_config_yml_file_path(cls) -> Path | None:
-        """Locate the `config.yml` file of the running connector."""
-        main_path = cls._get_connector_main_path()
-        config_yml_legacy_file_path = main_path.parent / "config.yml"
-        config_yml_file_path = main_path.parent.parent / "config.yml"
+    type: Literal["INTERNAL_ENRICHMENT"] = "INTERNAL_ENRICHMENT"
+    auto: bool = Field(
+        default=False,
+        description="Whether the connector should run automatically when an entity is created or updated.",
+    )
 
-        if config_yml_legacy_file_path.is_file():
-            return config_yml_legacy_file_path
-        elif config_yml_file_path.is_file():
-            return config_yml_file_path
-        return None
 
-    @classmethod
-    def _get_dot_env_file_path(cls) -> Path | None:
-        """Locate the `.env` file of the running connector."""
-        main_path = cls._get_connector_main_path()
-        dot_env_file_path = main_path.parent.parent / ".env"
+class BaseStreamConnectorConfig(_BaseConnectorConfig):
+    """Settings class for stream connectors.
 
-        return dot_env_file_path if dot_env_file_path.is_file() else None
+    Attributes:
+        type (str): The type of the connector, set to "STREAM" for stream connectors
+        live_stream_id (str): The ID of the live stream to connect to.
+        live_stream_listen_delete (bool): Whether to listen for delete events on the live stream.
+        live_stream_no_dependencies (bool): Whether to ignore dependencies when processing events from the live stream.
+    """
 
-    @classmethod
-    def settings_customise_sources(
-        cls,
-        settings_cls: type[BaseSettings],
-        init_settings: PydanticBaseSettingsSource,
-        env_settings: PydanticBaseSettingsSource,
-        dotenv_settings: PydanticBaseSettingsSource,
-        file_secret_settings: PydanticBaseSettingsSource,
-    ) -> tuple[PydanticBaseSettingsSource, ...]:
-        """Customise the sources of settings for the connector.
+    type: Literal["STREAM"] = "STREAM"
+    live_stream_id: str = Field(
+        description="The ID of the live stream to connect to.",
+    )
+    live_stream_listen_delete: bool = Field(
+        default=True,
+        description="Whether to listen for delete events on the live stream.",
+    )
+    live_stream_no_dependencies: bool = Field(
+        default=True,
+        description="Whether to ignore dependencies when processing events from the live stream.",
+    )
 
-        This method is called by the Pydantic BaseSettings class to determine the order of sources.
-        The configuration come in this order either from:
-            1. Environment variables
-            2. YAML file
-            3. .env file
-            4. Default values
 
-        The variables loading order will remain the same as in `pycti.get_config_variable()`:
-            1. If a config.yml file is found, the order will be: `ENV VAR` → config.yml → default value
-            2. If a .env file is found, the order will be: `ENV VAR` → .env → default value
-        """
-        config_yml_file_path = cls._get_config_yml_file_path()
-        if config_yml_file_path:
-            return (
-                env_settings,
-                YamlConfigSettingsSource(settings_cls, yaml_file=config_yml_file_path),
-            )
+class BaseInternalExportFileConnectorConfig(_BaseConnectorConfig):
+    """Settings class for internal export file connectors.
 
-        dot_env_file_path = cls._get_dot_env_file_path()
-        if dot_env_file_path:
-            return (
-                env_settings,
-                DotEnvSettingsSource(settings_cls, env_file=dot_env_file_path),
-            )
+    Attributes:
+        type (str): The type of the connector, set to "INTERNAL_EXPORT_FILE" for internal export file connectors.
+    """
 
-        return (env_settings,)
+    type: Literal["INTERNAL_EXPORT_FILE"] = "INTERNAL_EXPORT_FILE"
 
-    @classmethod
-    def build_loader_from_model(
-        cls, connector_settings: type["BaseConnectorSettings"]
-    ) -> type["_SettingsLoader"]:
-        """Build an untyped `_SettingsLoader` subclass for a connector's settings.
 
-        This method dynamically creates a subclass of `_SettingsLoader` that mirrors the
-        structure of the provided `BaseConnectorSettings` implementation. It disables all
-        Pydantic decoding, type coercion and validation so fields accept raw, unprocessed values.
+class BaseInternalImportFileConnectorConfig(_BaseConnectorConfig):
+    """Settings class for internal import file connectors.
 
-        The resulting model:
-        * Preserves values as-is from configuration sources
-        * Keeps YAML values as native Python types
-        * Keeps environment variables as plain strings
-        * Allows any field type (`Any`) without validation
+    Attributes:
+        type (str): The type of the connector, set to "INTERNAL_IMPORT_FILE" for internal import file connectors.
+        auto (bool): Whether the connector should run automatically when an entity is created or updated.
+    """
 
-        Args:
-            connector_settings (type[BaseConnectorSettings]): The typed connector settings class to mirror.
-
-        Returns:
-            type[_SettingsLoader]: A dynamically generated subclass of `_SettingsLoader`
-                where all fields accept raw, unvalidated input.
-        """
-
-        class SettingsLoader(_SettingsLoader): ...
-
-        model_fields = deepcopy(connector_settings.model_fields)
-        for field_info in model_fields.values():
-            annotation = field_info.annotation
-            if annotation and issubclass(annotation, BaseModel):
-                fields: dict[str, Any] = dict.fromkeys(
-                    annotation.model_fields.keys(), Any
-                )
-                untyped_model = create_model(
-                    f"{annotation.__name__}Untyped",
-                    __base__=annotation,
-                    **fields,
-                )
-                field_info.annotation = untyped_model
-                field_info.default_factory = untyped_model
-
-        SettingsLoader.model_fields = model_fields  # type: ignore
-        return SettingsLoader
+    type: Literal["INTERNAL_IMPORT_FILE"] = "INTERNAL_IMPORT_FILE"
+    auto: bool = Field(
+        default=False,
+        description="Whether the connector should run automatically when an entity is created or updated.",
+    )
 
 
 class BaseConnectorSettings(BaseConfigModel, ABC):
@@ -314,6 +253,11 @@ class BaseConnectorSettings(BaseConfigModel, ABC):
             super().__init__()
         except ValidationError as e:
             raise ConfigValidationError("Error validating configuration.") from e
+
+        self.logger.debug(
+            f"{self.__class__.__name__} instantiated succesfully",
+            {"settings": self.model_dump(mode="json")},
+        )
 
     @classmethod
     def config_json_schema(
@@ -477,6 +421,12 @@ class BaseConnectorSettings(BaseConfigModel, ABC):
 
         # Get config/env vars as dict to send for validation
         config_dict: dict[str, Any] = settings_loader().model_dump()
+
+        cls.logger.debug(
+            "Raw configuration data loaded for validation",
+            {"config_data": config_dict},
+        )
+
         return handler(config_dict)
 
     def to_helper_config(self) -> dict[str, Any]:
@@ -489,81 +439,3 @@ class BaseConnectorSettings(BaseConfigModel, ABC):
             # OpenCTIConnectorHelper handles missing fields with default values or internal logic.
             exclude_none=True,
         )
-
-
-class BaseExternalImportConnectorConfig(_BaseConnectorConfig):
-    """Settings class for external import connectors.
-
-    Attributes:
-        type (str): The type of the connector, set to "EXTERNAL_IMPORT" for external import connectors.
-        duration_period (timedelta): The period of time to await between two runs of the connector.
-    """
-
-    type: Literal["EXTERNAL_IMPORT"] = "EXTERNAL_IMPORT"
-    duration_period: timedelta = Field(
-        description="The period of time to await between two runs of the connector."
-    )
-
-
-class BaseInternalEnrichmentConnectorConfig(_BaseConnectorConfig):
-    """Settings class for internal enrichment connectors.
-
-    Attributes:
-        type (str): The type of the connector, set to "INTERNAL_ENRICHMENT" for internal enrichment connectors.
-        auto (bool): Whether the connector should run automatically when an entity is created or updated.
-    """
-
-    type: Literal["INTERNAL_ENRICHMENT"] = "INTERNAL_ENRICHMENT"
-    auto: bool = Field(
-        default=False,
-        description="Whether the connector should run automatically when an entity is created or updated.",
-    )
-
-
-class BaseStreamConnectorConfig(_BaseConnectorConfig):
-    """Settings class for stream connectors.
-
-    Attributes:
-        type (str): The type of the connector, set to "STREAM" for stream connectors
-        live_stream_id (str): The ID of the live stream to connect to.
-        live_stream_listen_delete (bool): Whether to listen for delete events on the live stream.
-        live_stream_no_dependencies (bool): Whether to ignore dependencies when processing events from the live stream.
-    """
-
-    type: Literal["STREAM"] = "STREAM"
-    live_stream_id: str = Field(
-        description="The ID of the live stream to connect to.",
-    )
-    live_stream_listen_delete: bool = Field(
-        default=True,
-        description="Whether to listen for delete events on the live stream.",
-    )
-    live_stream_no_dependencies: bool = Field(
-        default=True,
-        description="Whether to ignore dependencies when processing events from the live stream.",
-    )
-
-
-class BaseInternalExportFileConnectorConfig(_BaseConnectorConfig):
-    """Settings class for internal export file connectors.
-
-    Attributes:
-        type (str): The type of the connector, set to "INTERNAL_EXPORT_FILE" for internal export file connectors.
-    """
-
-    type: Literal["INTERNAL_EXPORT_FILE"] = "INTERNAL_EXPORT_FILE"
-
-
-class BaseInternalImportFileConnectorConfig(_BaseConnectorConfig):
-    """Settings class for internal import file connectors.
-
-    Attributes:
-        type (str): The type of the connector, set to "INTERNAL_IMPORT_FILE" for internal import file connectors.
-        auto (bool): Whether the connector should run automatically when an entity is created or updated.
-    """
-
-    type: Literal["INTERNAL_IMPORT_FILE"] = "INTERNAL_IMPORT_FILE"
-    auto: bool = Field(
-        default=False,
-        description="Whether the connector should run automatically when an entity is created or updated.",
-    )
